% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLargeScaleCharacteristics.R
\name{getLargeScaleCharacteristics}
\alias{getLargeScaleCharacteristics}
\title{Explain function}
\usage{
getLargeScaleCharacteristics(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  temporalWindows = list(c(NA, -366), c(-365, -91), c(-365, -31), c(-90, -1), c(-30,
    -1), c(0, 0), c(1, 30), c(1, 90), c(31, 365), c(91, 365), c(366, NA)),
  tablesToCharacterize = c("condition_occurrence", "drug_era", "procedure_occurrence",
    "measurement"),
  overlap = TRUE
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain the 'targetCohort' table and all the tables that we want to
characterize. It is a compulsory input, no default value is provided.}

\item{targetCohortName}{Name of the table in the cdm that contains the
target cohort that we want to characterize. It is a compulsory input, no
default value is provided.}

\item{targetCohortId}{Cohort definition id for the analyzed target cohorts.
It can be a vector or a number. If it is NULL all cohorts are analyzed. By
default: NULL.}

\item{temporalWindows}{Temporal windows that we want to characterize. It must
be a list of numeric vectors of length two. The tables will be characterized
between the first element and the second element respect to the
cohort_start_date of each individual. To refer to any time prior set NA the
first element of the vector. To refer to any time after the index date set NA
the second element of the vector. By default: list(c(NA, -366), c(-365, -90),
c(-365, -31), c(-90, -1), c(-30, -1), c(0, 0), c(1, 30), c(1, 90),
c(31, 365), c(91, 365), c(366, NA)).}

\item{tablesToCharacterize}{Name of the tables in the cdm that we want to
summarize. The available tables to characterize are: "visit_occurrence",
"condition_occurrence", "drug_exposure", "procedure_occurrence",
"device_exposure", "measurement", "observation", "drug_era", "condition_era"
and "specimen". By default: c("condition_occurrence", "drug_era",
"procedure_occurrence", "measurement").}

\item{overlap}{Whether you want to consider overlapping events (overlap =
TRUE) or only incident ones (overlap = FALSE).}
}
\value{
The output of this function is a table containing summary characteristics.
Key information like temporalWindows considered will be output in columns
window_id and window_name. tablesToChacaterize will be output in columns
table_id and table_name. Concept information is output in columns
concept_id, concept_name, concept_type and concept_count. Denominator information
denominator_count is also reported. "overlap" setting is also included
in the table for reference.
}
\description{
Explain function
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(LargeScaleCharacteristics)
cdm <- mockLargeScaleCharacteristics()
getLargeScaleCharacteristics(cdm, targetCohortName = c("cohort1"))
}
}
