% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LargeScaleCharacteristics-package.R
\docType{package}
\name{LargeScaleCharacteristics-package}
\alias{LargeScaleCharacteristics}
\alias{LargeScaleCharacteristics-package}
\title{LargeScaleCharacteristics: Large scale characterization using the OMOP Common Data Model}
\description{
This packages computes the large scale characterization of the OMOP Common Data Model at patient level or summarizing at cohort level.
}
\keyword{internal}
